move python version definitions
