+++ src/lightning/pytorch/trainer/connectors/signal_connector.py
@@ -11,8 +11,7 @@ from lightning_utilities.core.rank_zero import rank_prefixed_message
 
 import lightning.pytorch as pl
 from lightning.fabric.plugins.environments import SLURMEnvironment
-from lightning.fabric.utilities.imports import _IS_WINDOWS
-from lightning.pytorch.utilities.imports import _PYTHON_GREATER_EQUAL_3_8_0
+from lightning.fabric.utilities.imports import _IS_WINDOWS, _PYTHON_GREATER_EQUAL_3_8_0
 from lightning.pytorch.utilities.rank_zero import rank_zero_info
 
 # copied from signal.pyi
