+++ tests/tests_pytorch/trainer/logging_/test_eval_loop_logging.py
@@ -25,6 +25,7 @@ import pytest
 import torch
 from torch import Tensor
 
+from lightning.fabric.utilities.imports import _PYTHON_GREATER_EQUAL_3_8_0
 from lightning.pytorch import callbacks, Trainer
 from lightning.pytorch.callbacks.progress.rich_progress import _RICH_AVAILABLE
 from lightning.pytorch.demos.boring_classes import BoringModel, RandomDataset
@@ -32,7 +33,6 @@ from lightning.pytorch.loggers import TensorBoardLogger
 from lightning.pytorch.loops import _EvaluationLoop
 from lightning.pytorch.trainer.states import RunningStage
 from lightning.pytorch.utilities.exceptions import MisconfigurationException
-from lightning.pytorch.utilities.imports import _PYTHON_GREATER_EQUAL_3_8_0
 from tests_pytorch.helpers.runif import RunIf
 
 if _RICH_AVAILABLE:
